+++ internal/engine/util/bitmap.h
@@ -7,18 +7,20 @@
 
 #pragma once
 
+#include <stdlib.h>
+
 namespace bitmap {
 
 /* init a bitmap of which the total length is size */
-int create(char *&bitmap, int &bytes_count, int size);
+int create(char *&bitmap, int64_t &bytes_count, int64_t size);
 
 /* assume id not exceed the total size of bitmap */
-bool test(const char *bitmap, int id);
+bool test(const char *bitmap, int64_t id);
 
 /* assume id not exceed the total size of bitmap */
-void set(char *bitmap, int id);
+void set(char *bitmap, int64_t id);
 
 /* assume id not exceed the total size of bitmap */
-void unset(char *bitmap, int id);
+void unset(char *bitmap, int64_t id);
 
 }  // namespace bitmap
